Escena 1 sonorizada, más sonidos de impacto

+++ engine/engine/audio/sound_manager.h
@@ -94,7 +94,7 @@ public:
 	void activateSlowMo();
 	void desactivateSlowMo();
 
-	float getMaterialTagValue(std::string material);
+	int getMaterialTagValue(std::string material);
 
 	void ERRCHECK(FMOD_RESULT result);
